Correct import scope for redis module
Redis was being imported inside a function. References to redis outside
that function would make the program crash.

Now we try to import redis globally. If the import fails we just move
on. But later if we see that redis is enabled in the configuration we
check if redis is indeed defined. If not, the program exits. Otherwise
all is good.